requirements.txt
google-cloud-storage
Flask

